requirements-new.txt
aiohttp
async-timeout
backports.weakref
bleach
certifi
chardet
Django
gTTS
gTTS-token
html5lib
idna
Markdown
multidict
numpy
olefile
Pillow
protobuf
PyMarkov
requests
six
telepot
urllib3
Werkzeug
yarl
